requirements.txt
PyYAML
apache-airflow==1.10.4
flask_bcrypt